backend/requirements.txt
Flask==2.3.3
Flask-CORS==3.0.10
Flask-JWT-Extended
pymongo==4.1.1
python-dotenv
werkzeug
Flask-RESTful==0.3.9
pyjwt
